homework_2/requirements.txt
kafka-python-ng==2.2.3
pandas==2.2.0
openpyxl==3.1.2
langdetect==1.0.9
textblob==0.17.1
nltk==3.8.1
pyarrow==15.0.0